backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-decouple==3.8
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
django-extensions==3.2.3
